requirements.txt
opencv-python==4.10.0.84
pytesseract==0.3.13
googletrans==4.0.2
Pillow==11.1.0